PASSBOLT-2777 Remove the polyfill fetch from the list of dependencies for firefox only
PASSBOLT-2777 Add the original validator.js dependency, and extend it with isUtf8 & isUtf8Extended. Change the validation of the forms of the plugin setup, use the added utf8 rule instead of the legacy isAlphanumericSpecial rule

PASSBOLT-2777 Remove dev-dependencies from the repository. Most of the vendors are built in vendors.min.js

PASSBOLT-2777 Update package-lock.json. Only ASCII characters are now allowed for the visual security token