PB-8471 fix enabled plugin detection when a plugin is missing
When a plugin was missing the former code consider the plugin to be enabled.
This wasn't intended at the beginning as the check was always bypassed accidentaly.

Now when a plugin is non existing, it considers the plugin to be disabled.
When a plugiun exist, it has to explicitly define "enabled = false" to be considered
disabled. In other cases, the plugin is considered enabled.