PB-8371 provide dark theme for every iframes
The way it works is the following: a css is loaded for each theme with a media
attribute that ask to use them depending on the OS theme configuration. With
that approach the theme is dynamically chosen based on the configuration of the OS.

If the user has selected a specific theme (regardless of the OS config) then the
stylesheet.js will force the theme to be the right one.

PB-8371 improve stylesheet management by loading only necessary CSS